cannot find which part of Arduino code is wrong, fixed #1 by re-open serial port instead.

+++ Snail_Arduino/HandTieArduinoSystem/AD5231.cpp
@@ -1,6 +1,6 @@
 #include "AD5231.h"
 
-AD5231::AD5231(uint8_t slave_select){
+AD5231::AD5231(uint16_t slave_select){
 	SPI.begin();
 	SPI.setBitOrder(MSBFIRST);
